requirements.txt
torch
numpy
pandas
matplotlib
torchvision
tqdm
transformers
scikit-learn
Pillow
